Fix animation bug with adjustHandSpacing
Since the  tweens are getting places into a queue and the list of cards
to be tweened as being passed in as a reference, the amount of cards may
change before the tween is executed, leading to unexpected results.

The solutions is to pass the current list of cards that we mean to tween
as a copy, so that only the cards contained in the list at the time of
tween creation are animated.